Moved hook-specific stuff into their own package
