--- src/main/java/fr/estecka/variantscit/MixinGlobals.java
@@ -1,7 +1,0 @@
-package fr.estecka.variantscit;
-
-import fr.estecka.variantscit.trims.TrimPatternOverlay;
-
-public class MixinGlobals {
-	static public TrimPatternOverlay trimOverride = null;
-}
