+++ src/main/java/fr/estecka/variantscit/hooks/MixinGlobals.java
@@ -1,0 +1,5 @@
+package fr.estecka.variantscit.hooks;
+
+public class MixinGlobals {
+	static public TrimPatternOverlay trimOverride = null;
+}
