+++ src/main/java/fr/estecka/variantscit/mixin/EquipmentLayerRendererMixin.java
@@ -13,10 +13,10 @@ import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
-import fr.estecka.variantscit.MixinGlobals;
+import fr.estecka.variantscit.hooks.MixinGlobals;
+import fr.estecka.variantscit.hooks.TrimPatternOverlay;
 import fr.estecka.variantscit.VariantsCitMod;
 import fr.estecka.variantscit.reload.EModuleHook;
-import fr.estecka.variantscit.trims.TrimPatternOverlay;
 import net.minecraft.util.Util;
 import net.minecraft.client.renderer.entity.layers.EquipmentLayerRenderer;
 import net.minecraft.client.renderer.entity.layers.EquipmentLayerRenderer.TrimSpriteKey;
